oiqa_bpr_vmamba/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oiqa-bpr-vmamba"
version = "0.1.0"
description = "CVIQ-first reimplementation of a BPR + VMamba OIQA model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "scipy>=1.10",
  "scikit-learn>=1.3",
  "tqdm>=4.66",
  "opencv-python-headless>=4.9",
  "einops>=0.7",
  "timm>=1.0.8",
  "torch>=2.2",
]

[project.scripts]
oiqa-build-cviq-manifest = "oiqa_bpr_vmamba.cli.build_cviq_manifest:main"
oiqa-synthesize-degraded = "oiqa_bpr_vmamba.cli.synthesize_degraded_viewports:main"
oiqa-train-cviq = "oiqa_bpr_vmamba.cli.train_cviq:main"
oiqa-eval-cviq = "oiqa_bpr_vmamba.cli.eval_cviq:main"
oiqa-run-ablation = "oiqa_bpr_vmamba.cli.run_ablation:main"
oiqa-run-split-protocols = "oiqa_bpr_vmamba.cli.run_split_protocols:main"
oiqa-run-full-benchmark = "oiqa_bpr_vmamba.cli.run_full_benchmark:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
